Adjusts from Backend and FrontEnd
Backend:
- Insert send mail service;
- Added validation on endpoints;
- Correction of sql commands;

Frontend:
- Altered name from "yearMonth" to "monthYear"

+++ tcc-backend/src/main/java/com/tcc/professionalOfTheMonth/commons/GenericController.java
@@ -1,6 +1,8 @@
 package com.tcc.professionalOfTheMonth.commons;
 
 import lombok.Getter;
+
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -8,6 +10,7 @@ import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 import java.io.Serializable;
 import java.util.List;
+import java.util.Optional;
 
 public abstract class GenericController <E, ID extends Serializable, S extends GenericService<E, ID, ?>> {
 
@@ -25,23 +28,32 @@ public abstract class GenericController <E, ID extends Serializable, S extends G
 
     @GetMapping(path = "{id}")
     public ResponseEntity<E> findById(@PathVariable ID id) throws EntityNotFoundException {
-        return ResponseEntity.ok(service.findById(id)
-                .orElseThrow(EntityNotFoundException::new));
+    	
+    	Optional<E> opt = service.findById(id);
+    	
+    	if(opt.isPresent()) {
+    		return new ResponseEntity<>(opt.get(), HttpStatus.OK);
+    	}else {
+    		return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    	}
     }
 
     @PostMapping
-    public ResponseEntity<E> save(@RequestBody @Valid E body) {
-        return ResponseEntity.ok(service.save(body));
-    }
-
-    @PutMapping
-    public ResponseEntity<E> update(@RequestBody E body , @PathVariable ID id) {
+    public ResponseEntity<E> save(@Valid @RequestBody E body) {
         return ResponseEntity.ok(service.save(body));
     }
 
     @DeleteMapping(path = "{id}")
-    public void delete(@PathVariable ID id) {
-        service.deleteById(id);
+    public ResponseEntity<E> delete(@PathVariable ID id) {
+    	
+    	Optional<E> opt = service.findById(id);
+    	
+    	if(opt.isPresent()) {
+    		service.deleteById(id);
+    		return new ResponseEntity<>(opt.get(), HttpStatus.OK);
+    	}else {
+    		return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    	}
     }
 
 }
